rag/requirements.txt
# ─────────────────────────────────────────────
# requirements.txt（pip install -r requirements.txt）
# ─────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu          # 有 GPU 請改 faiss-gpu
sentence-transformers
transformers>=4.40.0
accelerate
torch
peft
pandas
numpy
opencc-python-reimplemented
scikit-learn
pymupdf